LayoutTests/http/tests/resourceLoadStatistics/telemetry-generation-advanced-functionality-database-expected.txt

Tests that advanced telemetry for prevalent resources in the SQLite Database is calculated correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Hosts classified correctly.
PASS testResult.numberOfPrevalentResources is 5
PASS testResult.numberOfPrevalentResourcesWithUserInteraction is 2
PASS testResult.numberOfPrevalentResourcesWithoutUserInteraction is 3
PASS testResult.topPrevalentResourceWithUserInteractionDaysSinceUserInteraction is 0
PASS testResult.medianDaysSinceUserInteractionPrevalentResourceWithUserInteraction is 1
PASS testResult.top3NumberOfPrevalentResourcesWithUI is 2
PASS testResult.top3MedianSubFrameWithoutUI is 5
PASS testResult.top3MedianSubResourceWithoutUI is 5
PASS testResult.top3MedianUniqueRedirectsWithoutUI is 4
PASS testResult.top3MedianDataRecordsRemovedWithoutUI is 0
PASS successfullyParsed is true

TEST COMPLETE
